requirements.txt
fastapi==0.111.0
uvicorn==0.26.1
SQLAlchemy==2.0.22
python-dotenv==1.0.1
passlib[argon2]==1.8.2
python-jose==3.3.0
